pylintrc: always ignore "too few public methods"
pylint thinks that no method on an object mean you better use a dict or
namedtuple.

No.

Basic config object subclassed in a custom config object with custom
attributes *and* also base attributes is an example.

And generally speaking, creating a methodless object involves more work
than just throwing a dict or namedtuple in the code, so it's highly
probable that the developer has taken the decision after having thought
about it.