test: raise error when using ineffective login()
when using login() or client_login(), we check if app is currently not
in 'NO_LOGIN' mode and if security service is running.

This is to prevent writing defective tests when using users; without
this SYSTEM user could be used during self.client requests, or security
would always say "yes" when checking roles and permissions.