disable csrf for form created outside app
specifically, pass "csrf_enabled=False" when parameter is not present,
no app context is set up (because if csrf is enabled flask-wtf will look
for current_app) and if there is a prefix, which is the case when the
form is created by de formfield.

When there is not prefix it's probably a "top-level" form, and we'd
better let application break so that developper knows that his/her
form is instanciated at runtime, not at app/request time.