Module json2: allow to call another search method
this is an api endpoint (Module.json2) that searches with a regular
database query (not Whoosh). Some modules may want to have there own
search method (so as to add bonus information into the text results for
example), so we call it if any or we rely on the basic
search (name.ilike).